第一次送样版本，睡眠时写flash（闪烁问题待解决）

+++ HARDWARE/task.h
@@ -29,7 +29,7 @@ typedef struct _xSys
 	uint8_t LowVoltageFlag;
 	
 	uint16_t EntreSleepTimeCount;  
-	uint8_t SleepTimeCount;  
+	uint16_t SleepTimeCount;  
 	
 	uint8_t IrWakeUPFlag;		
 	uint8_t ChargWakeUPFlag;
